make stack_visualization_data func, make config file and move args_parser func, change result file name
